Track who proposed a tentative match if manually entered.
This is not recorded permanently, but at least displayed on the tentative matches page.
Should prevent UserTwo from confirming a tentative match thinking there's a ton of evidence,
when in fact it has just been proposed by UserOne. Now UserTwo will be alerted that it's 
a manually entered match, and can take appropriate care.

As a consequence, self.tentativeMatches is now a list of dictionaries instead of a 
list of tuples.