Commented chemkin thermo file checking
The master RMG-Py checks chemkin thermo files for consistency, which
seems to throw a ChemkinError when trying to run several models with the importer.
Probably due to poorly formatted thermo files, but commented out the code for the time
being so the importer can run.

Removed duplicate code that came about during the rebase onto master.