PrIMe species writer for the chemkin importer
Takes the species data and writes the xml PrIMe species file.
At this stage, it does not print the PrIMe file, but the string is ready
to be written. Currently the method is called when a species match is
confirmed, but we should probably move this and call the method once
the entire model has been imported, then ask the user to confirm the writing
of the PrIMe files.

The PrIMe ID still needs to be sourced, and other inputs (e.g. copyright, source)
need to be added.